I've added geographical clustering data and a google map which displays squares around the clusters. Obviously a few of the clusters are too large to be useful. I have a square that goes from Arkansas to Redwood City. This is a result of the kmeans algorithm's preference for evenly sized groups over proximity. A simple solution would be to throw away groups that cover too wide an area, or perhaps groups which have the lowest checkins per sq-mile. I could also filter checkins by proximity to the user's current location. Users may also wish to limit to checkins made after a certain date. Adjusting the number of groups can have an impact, and I may way to give users control over this so they can help me find the ideal setting.
Though there is obviously much work to be done to the clustering to improve it there are still two things I need to turn my attention to; the first is combining two users' checkins which will yield clusters that fit both users (the point of the site), and the second is loading Venues for the clusters I've found. After that I need to enable, switching between timespans.

Finally, as the data gets bigger I need to start thinking about speeding it up. First I will move data loading to ajax calls. This will allow the interface to load immediately while the data loading spins. Second I need to start pre-processing checkin loads, and clustering in the TaskQueue. After all that I think it will be necessary to implement kmeans in MapReduce along with improvements to the centroid selection algorithm. Why is this so fun?